# swagger/linkrecommendations.yml
---
basePath: "/"
definitions:
  LinkRecommendationSet:
    type: object
    properties:
      links:
        type: array
        description: 'An array of link recommendation objects.'
        items:
          $ref: '#/definitions/LinkRecommendation'
      links_count:
        type: integer
        description: 'The number of link recommendations.'
      page_title:
        type: string
        description: 'The page title'
      pageid:
        type: integer
        description: 'The page ID'
      revid:
        type: integer
        description: 'The revision ID'
  LinkRecommendation:
    type: object
    properties:
      link_text:
        type: string
        description: 'The phrase to link in the article text.'
      wikitext_offset:
        type: integer
        description: 'The character offset describing where the anchor begins.'
      context_before:
        type: string
        description: 'The immediate preceding context for the anchor, in plaintext.'
      context_after:
        type: string
        description: 'The immediate succeeding context for the anchor, in plaintext.'
      link_target:
        type: string
        description: 'The article title that the anchor text should be linked to.'
      match_index:
        type: integer
        description: >
          The 0-based index of the link anchor within the list of matches when searching for the phrase to link
          within simple wikitext (top-level wikitext that's not part of any kind of wikitext construct).
      score:
        type: number
        description: 'The probability score that the link should be added.'
      link_index:
        type: integer
        description: "The 0-based index of the link recommendation within all link recommendations, sorted by wikitext offset."
info:
  description: See https://wikitech.wikimedia.org/wiki/Add_Link for more info
  title: Wikimedia Link Recommendation Service
  version: 1.0.0
paths:
  "/v0/linkrecommendations/{wiki_id}/{page_title}":
    post:
      x-monitor: true
      x-amples:
        - request:
            params:
              wiki_id: cswiki
              page_title: Lipsko
            headers:
              Content-Type: "application/json"
            body:
              pageid: 0
              revid: 0
              wikitext: "Foo"
            query:
              max_recommendations: 1
              threshold: 0.1
            response:
              status: 200
              body:
                links_count: 0
                page_title: "Lipsko"
                pageid: 0
                revid: 0
      parameters:
        - name: wiki_id
          type: string
          description: 'ID of the wiki for processing (e.g. "dewiki" or "simplewiki")'
          in: path
        - name: page_title
          type: string
          description: 'Page title of the article'
          in: path
        - name: threshold
          in: query
          required: false
          type: number
          default: 0.5
          description: 'Number between 0 and 1 for the threshold to use.'
        - name: max_recommendations
          in: query
          required: false
          type: number
          default: 15
          description: 'Maximum number of recommendations to return'
        - name: body
          in: body
          required: true
          schema:
            id: Input
            required:
              - wikitext
              - revid
              - pageid
            properties:
              wikitext:
                type: string
                description: 'Raw wikitext of the article'
              revid:
                type: integer
                description: 'Revision ID of the article'
              pageid:
                type: integer
                description: 'Page ID of the article'
    get:
      # Switch to true when we know that we can make requests to
      # cs.wikipedia.org/api.php via Envoy (T276217)
      x-monitor: false
      x-amples:
        - request:
            params:
              wiki_id: cswiki
              page_title: Lipsko
            query:
              max_recommendations: 1
              threshold: 0.1
            response:
              status: 200
              body:
                links_count: 1
                links:
                  link_index: 0
      parameters:
      - name: wiki_id
        in: path
        required: true
        type: string
        description: 'ID of the wiki for processing (e.g. "dewiki" or "simplewiki")'
      - name: page_title
        in: path
        required: true
        type: string
        description: 'Page title of the article'
      - name: threshold
        in: query
        required: false
        type: number
        default: 0.5
        description: 'Number between 0 and 1 for the threshold to use.'
      - name: max_recommendations
        in: query
        required: false
        type: number
        default: 15
        description: 'Maximum number of recommendations to return'
      responses:
        200:
          description: LinkRecommendationSet
          schema:
            $ref: '#/definitions/LinkRecommendationSet'
swagger: '2.0'